refactoring dashboard related components
