--- src/app/components/DashboardProjectList/index.ts
@@ -1,1 +1,0 @@
-export { default as DashboardProjectList } from './DashboardProjectList';
