--- src/app/components/ProjectListSkeleton/index.ts
@@ -1,1 +1,0 @@
-export { default as ProjectListSkeleton } from './ProjectListSkeleton';
