+++ src/app/components/dashboard-page/DashboardActions/CreateProjectButton/CreateProjectButton.tsx
@@ -3,7 +3,7 @@ import { useState, useCallback } from 'react';
 import Button from '@mui/material/Button';
 import AddIcon from '@mui/icons-material/Add';
 import { StyledModal } from '@/app/components/common/StyledModal';
-import { CreateProjectForm } from '@/app/components/CreateProjectForm';
+import { CreateProjectForm } from './CreateProjectForm';
 
 const CreateProjectButton = () => {
   const [isOpen, setIsOpen] = useState(false);
