+++ src/app/components/dashboard-page/DashboardActions/CreateProjectButton/CreateProjectForm/CreateProjectFormAlert/CreateProjectFormAlert.tsx
@@ -1,0 +1,18 @@
+import Box from '@mui/material/Box';
+import Alert from '@mui/material/Alert';
+
+const CreateProjectFormAlert = ({ errorMessage }: { errorMessage?: string }) => {
+  if (!errorMessage) {
+    return null;
+  }
+
+  return (
+    <Box marginBottom={2}>
+      <Alert variant="filled" icon={false} severity="error">
+        {errorMessage}
+      </Alert>
+    </Box>
+  );
+};
+
+export default CreateProjectFormAlert;
