+++ src/app/components/dashboard-page/DashboardActions/CreateProjectButton/CreateProjectForm/CreateProjectFormAlert/index.ts
@@ -1,0 +1,1 @@
+export { default as CreateProjectFormAlert } from './CreateProjectFormAlert';
