+++ src/app/components/dashboard-page/DashboardActions/CreateProjectButton/CreateProjectForm/CreateProjectFormBody/index.ts
@@ -1,0 +1,1 @@
+export { default as CreateProjectFormBody } from './CreateProjectFormBody';
