+++ src/app/components/dashboard-page/DashboardActions/CreateProjectButton/CreateProjectForm/CreateProjectFormHeader/index.ts
@@ -1,0 +1,1 @@
+export { default as CreateProjectFormHeader } from './CreateProjectFormHeader';
