+++ src/app/components/dashboard-page/DashboardActions/CreateProjectButton/CreateProjectForm/CreateProjectFormInputs/index.ts
@@ -1,0 +1,1 @@
+export { default as CreateProjectFormInputs } from './CreateProjectFormInputs';
