+++ src/app/components/dashboard-page/DashboardActions/index.ts
@@ -1,0 +1,1 @@
+export { default as DashboardActions } from './DashboardActions';
