+++ src/app/components/dashboard-page/DashboardProjects/DashboardProjectsHeader/index.ts
@@ -1,0 +1,1 @@
+export { default as DashboardProjectsHeader } from './DashboardProjectsHeader';
