+++ src/app/components/dashboard-page/DashboardProjects/DashboardProjectsTable/ProjectsNameFilter/index.ts
@@ -1,0 +1,1 @@
+export { default as ProjectsNameFilter } from './ProjectsNameFilter';
