+++ src/app/components/dashboard-page/DashboardProjects/DashboardProjectsTable/ProjectsPagination/index.ts
@@ -1,0 +1,1 @@
+export { default as ProjectsPagination } from './ProjectsPagination';
