+++ src/app/components/dashboard-page/DashboardProjects/DashboardProjectsTable/ProjectsTable/index.ts
@@ -1,0 +1,1 @@
+export { default as ProjectsTable } from './ProjectsTable';
