+++ src/app/components/dashboard-page/DashboardProjects/DashboardProjectsTable/index.ts
@@ -1,0 +1,1 @@
+export { default as DashboardProjectsTable } from './DashboardProjectsTable';
