+++ src/app/components/dashboard-page/DashboardProjects/index.ts
@@ -1,0 +1,1 @@
+export { default as DashboardProjects } from './DashboardProjects';
