+++ src/app/components/dashboard-page/index.ts
@@ -1,0 +1,1 @@
+export { default as DashboardContent } from './DashboardContent';
