+++ src/app/components/skeletons/DashboardProjectsTableSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { default as DashboardProjectsTableSkeleton } from './DashboardProjectsTableSkeleton';
